refactor: separated save plan logic
